requirements.txt
requests
json
os
botometer
google-cloud-language